rearrange set and samples
